Split code into 3 classes
This is just the three classes - behavior, brain, and gut.
To run it, you need to create one object each. The brain object takes the behavior object as input.
Then, you use the stateSet[currentStateIndex] from brain to get current state and input it to the gut.
Use the gut output to update the reward in the brain. At this point you can also store the state and action that has just occured.
Then use the brain's learnAndAct() to generate action and new state and update qTable. You can then input stateSet[currentStateIndex]  to the gut again and so on.